monitoring, e2e, Move endpoints test to monitoring lane
This commit syntactically moves the test as is to the monitoring lane
Additionally, it adds an AfterEach to make sure the CR is removed after
each test
